shared/icon-label: Prevent moc from being appended to global file
Like described in 5eb04cdf9dc29a8ee9238da3a2e46dbfacacf9ec, this
prevents the "moc_icon-label.cpp" from being included in the global
"mocs_compilation.cpp" file. Looks a bit ridiculous to create a new file
for this, but that's just because content-wise icon-label is only a
header.